+++ swiyu-sdjwt-verifier/src/main/java/ch/admin/bj/swiyu/sdjwtverifier/exception/SdJwtVerificationException.java
@@ -1,0 +1,18 @@
+package ch.admin.bj.swiyu.sdjwtverifier.exception;
+
+public class SdJwtVerificationException extends Exception {
+    private static final long serialVersionUID = 1l;
+
+    public SdJwtVerificationException(String message) {
+        super(message);
+    }
+
+    public SdJwtVerificationException(Throwable cause) {
+        super(cause);
+    }
+
+    public SdJwtVerificationException(String message, Throwable cause) {
+        super(message, cause);
+    }
+    
+}